% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{The class linreg is similar to lm function in R which is used for linear modelling.}
\arguments{
\item{formula}{Formula used for calculation of dependent factor.}

\item{data}{is a Data frame.}
}
\value{
linear model based on \code{formula} and \code{data}
}
\description{
The class linreg is similar to lm function in R which is used for linear modelling.
}
\details{
This class use RC Class in order to make Object oriented programming.

We are here making linear model custom way which is based on lm function which already exit in R. 

Please refer below doc to read more on lm function
\url{https://www.rdocumentation.org/packages/stats/versions/3.6.0/topics/lm}
}

\examples{
\dontrun{
a = linreg(formula,data)
a$print()
a$plot()
a$summary()
}
}
